+++ container.php
@@ -1,13 +1,10 @@
 <?php
 
-/*
-This file returns the global DI container for the application. The DI container
-which gets returned must be compatible with the symfony DI container interface.
-If you want load an different configuration depending on the environment you can 
-change the "config.file" parameter.
-*/
-
-$container = new Fusio\Impl\Dependency\Container();
-$container->setParameter('config.file', __DIR__ . '/configuration.php');
-
-return $container;
+return \PSX\Framework\Dependency\ContainerBuilder::build(
+    __DIR__,
+    true,
+    __DIR__ . '/vendor/psx/framework/resources/container.php',
+    __DIR__ . '/vendor/fusio/impl/resources/container.php',
+    __DIR__ . '/resources/container.php',
+    ...\Fusio\Impl\Adapter\AdapterFinder::getFiles(__DIR__ . '/provider.php')
+);